feat: api request in transition page
