+++ src/api/transactions/service/index.ts
@@ -1,15 +1,11 @@
 import { updateTransaction } from './updateTransaction.service';
 import { findOneTransaction } from './findOneTransaction.service';
-import {
-  listTransactionsByParams,
-  loadTotalizers,
-} from './listTransactionsByParams.service';
+import { listTransactionsByParams } from './listTransactionsByParams.service';
 import { createTransaction } from './createTransaction.service';
 
 export const transactionsService = {
   listTransactionsByParams,
   createTransaction,
-  loadTotalizers,
   findOneTransaction,
   updateTransaction,
 };
