+++ src/api/transactions/service/listTransactionsByParams.service.ts
@@ -1,42 +1,21 @@
-import { ResponseDefault } from './../../../interface/responseDefault.interface';
-import { queryStringBuilder } from '@/utils/queryStringBuilder';
+import { useUser } from '@/hooks/useUser';
 import api from '../../../services/api';
 import {
   RequestTransactions,
-  ResponseTotalizer,
   ResponseTransactionsByParams,
 } from '../interface/listTransactionByParamns';
 
 export const listTransactionsByParams = ({
-  limit,
-  page,
-  userId,
-  query: { categoryId, date, type },
+  endDate,
+  startDate,
+  token,
 }: RequestTransactions) => {
-  // return api.get<ResponseTransactionsByParams>(
-  //   `transactions/user/${userId}/all?limit=${limit}&page=${page}`,
-  // );
-
-  const query: { [key: string]: any } = {};
-  if (categoryId) query.categoryId = categoryId;
-  if (date) query.date = date;
-  if (type) query.type = type;
-
-  const queryString = queryStringBuilder(query);
-
   return api.get<ResponseTransactionsByParams>(
-    `transactions/user/${userId}?${queryString}`,
-  );
-};
-
-export const loadTotalizers = ({ userId, query }: RequestTransactions) => {
-  // return api.get<ResponseTransactionsByParams>(
-  //   `transactions/user/${userId}/all?limit=${limit}&page=${page}`,
-  // );
-
-  const queryString = queryStringBuilder(query);
-
-  return api.get<ResponseTotalizer>(
-    `transactions/user/${userId}/totalizer?${queryString}`,
+    `transactions/list?${startDate}&?${endDate}`,
+    {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    },
   );
 };
